refactor: declare user and account types once in contracts
`User`, `Account` and `AccountSettings` were each declared twice — in
`packages/data/src/users/data.ts` and again under `apps/web/src` — on
either side of the RPC boundary, and had begun to disagree: the server
typed `Account.email` as `string` while the client typed it optional.
`UserGroupReference` and `UserOption` were a third and fourth spelling
of `GroupMinimal` and `UserNested`, and `UserSearchResults` re-spelled
`SearchResult`'s five fields inline.

Move `User`, `Account`, `AccountSettings`, `QuickAnalyzeWorkflow` and
`UserSearchResult` into `@virtool/contracts`; delete
`apps/web/src/users/types.ts` and `apps/web/src/account/types.ts`. The
data layer keeps what only it uses — `StoredAccountSettings` and its
mappers, `DEFAULT_USER_SETTINGS`, and its `*Values` argument types.

`email` resolves to the server's reading: the column is `NOT NULL`
defaulting to `""`, and clearing an address writes the empty string, so
it is never absent. `AccountEmail` takes it as required.
