+++ apps/web/src/administration/utils.ts
@@ -1,5 +1,4 @@
-import type { Account } from "@account/types";
-import type { Permission } from "@virtool/contracts";
+import type { Account, Permission } from "@virtool/contracts";
 import {
 	AdministratorPermissions,
 	hasSufficientAdminRole,
