+++ apps/web/src/tests/fake/account.ts
@@ -1,6 +1,5 @@
-import type { Account, AccountSettings } from "@account/types";
 import { faker } from "@faker-js/faker";
-import type { ApiKey } from "@virtool/contracts";
+import type { Account, AccountSettings, ApiKey } from "@virtool/contracts";
 import { createFakePermissions } from "./permissions";
 import { createFakeUser } from "./user";
