+++ apps/web/src/tests/fake/user.ts
@@ -1,6 +1,5 @@
 import { faker } from "@faker-js/faker";
-import type { User } from "@users/types";
-import type { UserNested } from "@virtool/contracts";
+import type { User, UserNested } from "@virtool/contracts";
 import { createFakeGroupMinimal } from "./groups";
 import { createFakePermissions } from "./permissions";
