+++ apps/web/src/tests/setup.tsx
@@ -1,5 +1,4 @@
 import { accountQueryKeys } from "@account/keys";
-import type { Account } from "@account/types";
 import { faker } from "@faker-js/faker";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import {
@@ -10,6 +9,7 @@ import {
 	Outlet,
 	RouterProvider,
 } from "@tanstack/react-router";
+import type { Account } from "@virtool/contracts";
 import { rootQueryKeys } from "@wall/keys";
 import "@testing-library/jest-dom/vitest";
 import {
